move tests into SampleTest file
